+++ errors/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## To be Released
 
+## v2.5.1
+
+* chore(go): corrective bump - Go version regression from 1.24.3 to 1.24
+
 * fix(errors): `Build` returns `error` not `*ValidationErrors` (BREAKING CHANGE)
 * refactor(errors): remove deprecated functions (BREAKING CHANGE)